add featues
